Move fonts to its own module, add stream defaults
Move all font family/colours to its own model.
Add default values for generating streams.
